+++ src/data/arbitrary/cigar.rs
@@ -1,7 +1,10 @@
 use super::impl_deref;
-use crate::data::cigar::{Cigar, Ciglet};
+use crate::{
+    alignment::AlignmentStates,
+    data::cigar::{Cigar, Ciglet},
+};
 use arbitrary::{Arbitrary, Result, Unstructured};
-use std::num::NonZeroUsize;
+use std::{marker::PhantomData, num::NonZeroUsize};
 
 impl<'a> Arbitrary<'a> for Ciglet {
     #[inline]
@@ -21,6 +24,13 @@ pub struct CigletOp(pub Ciglet);
 
 impl_deref! {CigletOp, Ciglet}
 
+impl From<CigletOp> for Ciglet {
+    #[inline]
+    fn from(value: CigletOp) -> Self {
+        value.0
+    }
+}
+
 impl<'a> Arbitrary<'a> for CigletOp {
     #[inline]
     fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
@@ -39,41 +49,51 @@ impl<'a> Arbitrary<'a> for Cigar {
     }
 }
 
-/// A wrapper around [`Cigar`] such that the implementation of
-/// [`Arbitrary`](https://docs.rs/arbitrary/latest/arbitrary/trait.Arbitrary.html)
-/// only generates Cigar string with the following assumptions:
-/// * The string consists of alternating numbers and operations
-/// * The numbers must be between 1 and `usize::MAX`
-/// * The operations must be in MIDNSHP=X
-#[derive(Debug)]
-pub struct CigarIncOp(pub Cigar);
-
-impl_deref! {CigarIncOp, Cigar}
-
-impl<'a> Arbitrary<'a> for CigarIncOp {
+impl<'a> Arbitrary<'a> for AlignmentStates {
+    #[inline]
     fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
-        let mut out = Vec::new();
-        for ciglet in u.arbitrary_iter::<CigletOp>()?.flatten() {
-            out.extend_from_slice(ciglet.inc.to_string().as_bytes());
-            out.push(ciglet.op);
-        }
-        Ok(CigarIncOp(Cigar(out)))
+        Ok(AlignmentStates(Vec::<Ciglet>::arbitrary(u)?))
     }
 }
 
 /// A wrapper around [`Cigar`] such that the implementation of
 /// [`Arbitrary`](https://docs.rs/arbitrary/latest/arbitrary/trait.Arbitrary.html)
-/// is similar to [`CigarIncOp`] but may contain leading zeros before numbers.
+/// is generated from [`Ciglet`] structs.
+///
+/// The type of the ciglet is specified with `C`, which could be [`Ciglet`] or
+/// [`CigletOp`]. If `Z` is true, then leading zeros are added to the start of
+/// some increments in the CIGAR string.
 #[derive(Debug)]
-pub struct CigarIncOpLeadingZero(pub Cigar);
+pub struct CigarArbitrary<C, const Z: bool>(pub Cigar, PhantomData<C>);
 
-impl_deref! {CigarIncOpLeadingZero, Cigar}
+impl_deref! {CigarArbitrary<C, Z>, Cigar, <C, const Z: bool>}
 
-impl<'a> Arbitrary<'a> for CigarIncOpLeadingZero {
+impl<'a, C> Arbitrary<'a> for CigarArbitrary<C, false>
+where
+    C: Arbitrary<'a> + Into<Ciglet>,
+{
     fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
-        struct CigletOpWithLen(CigletOp, usize);
+        let mut out = Vec::new();
+        for ciglet in u.arbitrary_iter::<C>()?.flatten() {
+            let ciglet = ciglet.into();
+            out.extend_from_slice(ciglet.inc.to_string().as_bytes());
+            out.push(ciglet.op);
+        }
+        Ok(CigarArbitrary(Cigar(out), PhantomData))
+    }
+}
 
-        impl<'a> Arbitrary<'a> for CigletOpWithLen {
+impl<'a, C> Arbitrary<'a> for CigarArbitrary<C, true>
+where
+    C: Arbitrary<'a> + Into<Ciglet>,
+{
+    fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
+        struct CigletOpWithLen<C>(C, usize);
+
+        impl<'a, C> Arbitrary<'a> for CigletOpWithLen<C>
+        where
+            C: Arbitrary<'a> + Into<Ciglet>,
+        {
             fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
                 let ciglet_op = u.arbitrary()?;
                 let len = u.arbitrary_len::<u8>()?;
@@ -82,11 +102,59 @@ impl<'a> Arbitrary<'a> for CigarIncOpLeadingZero {
         }
 
         let mut out = Vec::new();
-        for CigletOpWithLen(ciglet, leading_zeros) in u.arbitrary_iter::<CigletOpWithLen>()?.flatten() {
+        for CigletOpWithLen(ciglet, leading_zeros) in u.arbitrary_iter::<CigletOpWithLen<C>>()?.flatten() {
             out.extend(std::iter::repeat_n(b'0', leading_zeros));
+            let ciglet = ciglet.into();
             out.extend_from_slice(ciglet.inc.to_string().as_bytes());
             out.push(ciglet.op);
         }
-        Ok(CigarIncOpLeadingZero(Cigar(out)))
+        Ok(CigarArbitrary(Cigar(out), PhantomData))
+    }
+}
+
+/// A wrapper around [`AlignmentStates`] offering more flexibility on the
+/// [`Arbitrary`] implementation.
+///
+/// ## Parameters:
+///
+/// * `C`: The [`Ciglet`] type or arbitrary wrapper for generating the ciglets
+/// * `D`: If true, ensures adjacent operations are different
+#[derive(Debug)]
+pub struct AlignmentStatesArbitrary<C, const D: bool>(pub AlignmentStates, PhantomData<C>);
+
+impl_deref! {AlignmentStatesArbitrary<C, D>, AlignmentStates, <C, const D: bool>}
+
+impl<'a, C, const D: bool> Arbitrary<'a> for AlignmentStatesArbitrary<C, D>
+where
+    C: Arbitrary<'a> + Into<Ciglet>,
+{
+    fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
+        let mut vec = Vec::<C>::arbitrary(u)?.into_iter().map(Into::into).collect::<Vec<_>>();
+
+        // Adjust the operations so that adjacent ones are not the same, if D is
+        // true
+        if let Some(mut last_ciglet) = vec.first().copied()
+            && D
+        {
+            for next_ciglet in &mut vec[1..] {
+                if next_ciglet.op == last_ciglet.op {
+                    next_ciglet.op = match last_ciglet.op {
+                        b'M' => b'I',
+                        b'I' => b'D',
+                        b'D' => b'N',
+                        b'N' => b'S',
+                        b'S' => b'H',
+                        b'H' => b'P',
+                        b'P' => b'=',
+                        b'=' => b'X',
+                        b'X' => b'M',
+                        _ => unreachable!(),
+                    }
+                }
+                last_ciglet = *next_ciglet;
+            }
+        }
+
+        Ok(AlignmentStatesArbitrary(AlignmentStates(vec), PhantomData))
     }
 }